Resolve cyclic imports
1. Move the Task structs from recommendation to config
2. Also, the tasks and rules are config specific not recommendation specific
3. Remove the import of config from Recommendation
4. Updated some log statements
